separate edit user role from new user role
